pyproject.toml
[tool.poetry]
name = "prom-otel"
version = "0.1.0"
description = ""
authors = ["isabellabissoli <78965606+isabellabissoli@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-distro = "^0.50b0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-api = "^1.29.0"
opentelemetry-instrumentation = "^0.50b0"
opentelemetry-exporter-otlp = "^1.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
